WIP Added std::ref() to the registers - added rst instruction - added bunch of debug printing

+++ src/CPU/Implementations/JUMP.cpp
@@ -19,7 +19,7 @@ uint8_t	CPU::JUMP_R16(uint16_t &reg)
 
 uint8_t		CPU::JUMP_RELATIVE_IM8()
 {
-	PC.setValue(PC.getValue() + fetchImmediateByte());
+	PC.setValue(PC.getValue() + static_cast<int8_t>(fetchImmediateByte()));
 	return (1);
 }
 
@@ -27,13 +27,13 @@ uint8_t		CPU::JUMP_RELATIVE_FLAG_IM8(Flag flag)
 {
 	uint8_t	data = fetchImmediateByte();
 	if (checkFlag(flag))
-		PC.setValue(PC.getValue() + data);
+		PC.setValue(PC.getValue() + static_cast<int8_t>(data));
 	return (1);
 }
 uint8_t		CPU::JUMP_RELATIVE_NFLAG_IM8(Flag flag)
 {
 	uint8_t	data = fetchImmediateByte();
 	if (!checkFlag(flag))
-		PC.setValue(PC.getValue() + data);
+		PC.setValue(PC.getValue() + static_cast<int8_t>(data));
 	return (1);
 }
